Move the webserver code to its own module
