+++ now_playing.py
@@ -10,15 +10,11 @@
 import logging
 import logging.handlers
 import os
-from http.server import HTTPServer, BaseHTTPRequestHandler
 import pathlib
 import shutil
-from socketserver import ThreadingMixIn
 import sys
-import tempfile
 import threading
 import time
-import urllib.parse
 
 from PyQt5.QtCore import \
                             pyqtSignal, \
@@ -38,6 +34,7 @@ import nowplaying.serato
 import nowplaying.settingsui
 import nowplaying.utils
 import nowplaying.version
+import nowplaying.webserver
 
 QAPP = QApplication(sys.argv)
 QAPP.setOrganizationName('com.github.em1ran')
@@ -131,7 +128,7 @@ class Tray:  # pylint: disable=too-many-instance-attributes
         self.trackthread.start()
 
         # Start the webserver
-        self.webthread = WebServer()
+        self.webthread = nowplaying.webserver.WebServer(config=CONFIG)
         self.webthread.webenable[bool].connect(self.webenable)
         self.webthread.start()
 
@@ -208,216 +205,6 @@ class Tray:  # pylint: disable=too-many-instance-attributes
             nowplaying.utils.writetxttrack(filename=CONFIG.file, clear=True)
         # calling exit should call __del__ on all of our QThreads
         QAPP.exit(0)
-
-
-class WebHandler(BaseHTTPRequestHandler):
-    ''' Custom handler for built-in webserver '''
-    def do_GET(self):  # pylint: disable=invalid-name
-        '''
-            HTTP GET
-                - if there is an index.htm file to read, give it out
-                  then delete it
-                - if not, have our reader check back in 5 seconds
-
-            Note that there is very specific path information
-            handling here.  So any chdir() that happens MUST
-            be this directory.
-
-            Also, doing it this way means the webserver can only ever
-            share specific content.
-        '''
-
-        global CONFIG
-
-        # see what was asked for
-        parsedrequest = urllib.parse.urlparse(self.path)
-
-        if parsedrequest.path in ['/favicon.ico']:
-            self.send_response(200)
-            self.send_header('Content-type', 'image/x-icon')
-            self.end_headers()
-            with open(CONFIG.iconfile, 'rb') as iconfh:
-                self.wfile.write(iconfh.read())
-            return
-
-        if parsedrequest.path in ['/', 'index.html', 'index.htm']:
-            self.send_response(200)
-            self.send_header('Content-type', 'text/html')
-            self.end_headers()
-            if os.path.isfile('index.htm'):
-                with open('index.htm', 'rb') as indexfh:
-                    self.wfile.write(indexfh.read())
-                os.unlink('index.htm')
-                return
-
-            self.wfile.write(b'<!doctype html><html lang="en">')
-            self.wfile.write(
-                b'<head><meta http-equiv="refresh" content="5" ></head>')
-            self.wfile.write(b'<body></body></html>\n')
-            return
-
-        if parsedrequest.path in ['/index.txt']:
-            self.send_response(200)
-            self.send_header('Content-type', 'text/plain')
-            self.end_headers()
-            with open(CONFIG.file, 'rb') as textfh:
-                self.wfile.write(textfh.read())
-            return
-
-        if parsedrequest.path in ['/cover.jpg']:
-            if os.path.isfile('cover.jpg'):
-                self.send_response(200, 'OK')
-                self.send_header('Content-type', 'image/jpeg')
-                self.end_headers()
-                with open('cover.jpg', 'rb') as indexfh:
-                    self.wfile.write(indexfh.read())
-                return
-
-        if parsedrequest.path in ['/cover.png']:
-            if os.path.isfile('cover.png'):
-                self.send_response(200, 'OK')
-                self.send_header('Content-type', 'image/png')
-                self.end_headers()
-                with open('cover.png', 'rb') as indexfh:
-                    self.wfile.write(indexfh.read())
-                return
-
-        self.send_error(404)
-
-    def log_message(self, format, *args):  ## pylint: disable=redefined-builtin
-        logging.info("%s - - [%s] %s\n", self.address_string(),
-                     self.log_date_time_string(), format % args)
-
-
-class ThreadingWebServer(ThreadingMixIn, HTTPServer):
-    ''' threaded webserver object '''
-    daemon_threads = True
-    allow_reuse_address = True
-
-
-class WebServer(QThread):
-    ''' Now Playing built-in web server using custom handler '''
-
-    webenable = pyqtSignal(bool)
-
-    def __init__(self, parent=None):
-        QThread.__init__(self, parent)
-        self.server = None
-        self.prevport = 0
-        self.prevdir = None
-        self.endthread = False
-        self.setObjectName('WebServerThread')
-
-    def run(self):  # pylint: disable=too-many-branches, too-many-statements
-        '''
-            Configure a webserver.
-
-            The sleeps are here to make sure we don't
-            tie up a CPU constantly checking on
-            status.  If we cannot open the port or
-            some other situation, we bring everything
-            to a halt by triggering pause.
-
-            But in general:
-
-                - web server thread starts
-                - check if web serving is running
-                - if so, open ANOTHER thread (MixIn) that will
-                  serve connections concurrently
-                - if the settings change, then another thread
-                  will call into this one via stop() to
-                  shutdown the (blocking) serve_forever()
-                - after serve_forever, effectively restart
-                  the loop, checking what values changed, and
-                  doing whatever is necessary
-                - land back into serve_forever
-                - rinse/repeat
-
-        '''
-        global CONFIG
-
-        threading.current_thread().name = 'WebServer'
-        while not CONFIG.httpenabled and not self.endthread and not CONFIG.initialized:
-            time.sleep(5)
-            CONFIG.get()
-
-        logging.debug('Web server thread started')
-
-        while not self.endthread:
-            logging.debug('Starting main loop')
-            while CONFIG.paused or not CONFIG.httpenabled:
-                time.sleep(5)
-                CONFIG.get()
-
-            resetserver = False
-            CONFIG.get()
-
-            if CONFIG.usinghttpdir:
-                if CONFIG.usinghttpdir != self.prevdir:
-                    logging.info('Using web server dir %s',
-                                 CONFIG.usinghttpdir)
-                    self.prevdir = CONFIG.usinghttpdir
-                    resetserver = True
-            else:
-                if not self.prevdir:
-                    logging.debug('No web server dir?!?')
-                    self.prevdir = tempfile.gettempdir()
-                    CONFIG.setusinghttpdir(self.prevdir)
-                    logging.info('Using web server dir %s', self.prevdir)
-
-            if not os.path.exists(self.prevdir):
-                try:
-                    pathlib.Path(self.prevdir).mkdir(parents=True,
-                                                     exist_ok=True)
-                except Exception as error:  # pylint: disable=broad-except
-                    logging.error('Web server threw exception! %s', error)
-                    self.webenable.emit(False)
-
-            os.chdir(self.prevdir)
-
-            if CONFIG.httpport != self.prevport:
-                if self.prevport:
-                    resetserver = True
-                self.prevport = CONFIG.httpport
-
-            if resetserver:
-                logging.debug('asked to reset')
-                self.stop()
-                time.sleep(5)
-
-            try:
-                self.server = ThreadingWebServer(('0.0.0.0', CONFIG.httpport),
-                                                 WebHandler)
-            except Exception as error:  # pylint: disable=broad-except
-                logging.error(
-                    'Web server threw exception on thread create: %s', error)
-                self.webenable.emit(False)
-
-            try:
-                if self.server:
-                    self.server.serve_forever()
-            except KeyboardInterrupt:
-                pass
-            except Exception as error:  # pylint: disable=broad-except
-                logging.error('Web server threw exception after forever: %s',
-                              error)
-            finally:
-                if self.server:
-                    self.server.shutdown()
-
-    def stop(self):
-        ''' method to stop the thread '''
-        logging.debug('WebServer asked to stop')
-        if self.server:
-            self.server.shutdown()
-            self.server.server_close()
-
-    def __del__(self):
-        logging.debug('Web server thread is being killed!')
-        self.endthread = True
-        self.stop()
-        self.wait()
-
 
 class TrackPoll(QThread):
     '''
