+++ nowplaying/version.py
